Fix Cloud#306 by moving `resultFuture.complete` to the `commandConsumer`.
Previously, the `resultFuture` was being `complete`d immediately after `commandConsumer` was invoked, and since `commandConsumer` runs in a non-blocking manner, it swallowed the exception.
Therefore, `resultFuture.complete` was moved into the `commandConsumer`, specifically within `whenComplete`. That way, it would only be completed when the execution has been finished.

There was also some other minor refactors, so that there's only one `return resultFuture`, which occurs at the end of the method.