feature: tooltip support
The brigadier manager supports tooltips and allows for the registration of T->Message mappings.

cloud-paper makes use of the completion API and adds the ability to register T->Component mappings.

Velocity has a native Component->Message mapping so it will make use of that one.

I opted to rename the suggestion method from "simple" to "of" and added a static-import-friendly "suggestion" method alongside it.