feat: allow for generic suggestions
This forces all suggestions to go through the suggestion factory (previously engine). This means that we do not have to pollute the entire codebase with yet another generic parameter (CommandManager is used in MANY places) that we generally do not need to care about.

The platforms can override the suggestion factory method and work with a sub-type of suggestion.

This is in preparation for tooltip support (see #66)