feat: add CommandDescription
A command description is a pair of argument descriptions (a short description and a verbose description).

This uses argument descriptions as they are already well-supported by the platforms. This means that Minecraft commands can use the rich descriptions and all the features that come with that.

The old meta keys have been removed. A separate PR will address the `@CommandDescription` annotation.

This should solve #527.