chore: clean up CommandMeta
This PR removes `CommandMeta.Key` and replaces all usage of it with `CloudKey`. It existed because of legacy reasons, and we no longer have a use for it.

SimpleCommandMeta no longer has to be referenced directly, and the command meta builder has been extracted and no longer exposes the implementation. The class remains public and non-final in case anyone wants to extend it.

The `CommandManager#createDefaultCommandMeta` method has a default implementation that returns an empty instance. It is very rare for the implementation of this to do anything else anyway.