feat(core): make `CommandComponent` immutable
The "owning command" has been moved into the command node, as the node is always present when we access the command anyway. This allows us to only have final fields in the command component.

This also means that we no longer need to copy the components before storing them in the command tree, and we can instead re-use the components.